기초 문제

+++ 코드업_파이썬 기초문제/6035_실수 2개 입력받아 곱 계산하기.py
@@ -1,4 +1,4 @@
 # 6035 실수 2개 입력받아 곱 계산하기
-n1, n2 = map(int,input().split())
+n1, n2 = map(float,input().split())
 multi = n1 * n2
 print(multi)
